Use num-traits for the trait
