+++ Cargo.toml
@@ -13,6 +13,7 @@ documentation = "https://docs.rs/lambert_w"
 
 [dependencies]
 libm = { version = "0.2", optional = true }
+num-traits = { version = "0.2.19", default-features = false }
 
 [dev-dependencies]
 approx = { version = "0.5.1", default-features = false }
@@ -25,10 +26,10 @@ no-panic = "0.1"
 default = ["libm"]
 # If the `std` feature is disabled, this feature uses the [`libm`](https://crates.io/crates/libm) crate
 # to compute square roots and logarithms during function evaluation instead of the standard library.
-libm = ["dep:libm"]
+libm = ["dep:libm", "num-traits/libm"]
 # Use the standard library to compute square roots and logarithms for a potential performance gain.
 # When this feature is disabled the crate is `no_std` compatible.
-std = []
+std = ["num-traits/std"]
 
 [package.metadata.docs.rs]
 # Document all features.
