+++ src/dw0c.rs
@@ -1,6 +1,6 @@
 use crate::{
     elementary::{ln, sqrt},
-    rational::rational_function,
+    rational_function,
 };
 
 // This is an implementation of the approximation of the principal
