+++ src/lib.rs
@@ -11,7 +11,6 @@ compile_error!("at least one of the `std` or `libm` features must be enabled");
 mod dw0c;
 mod dwm1c;
 mod elementary;
-mod rational;
 mod sw0;
 mod sw0f;
 mod swm1;
@@ -295,4 +294,25 @@ impl LambertW for f64 {
     fn lambert_wm1(self) -> Self::Output {
         lambert_wm1(self)
     }
+}
+
+/// Evaluate a rational function at `x` using Horner's method.
+#[inline(always)]
+#[cfg_attr(all(test, assert_no_panic), no_panic::no_panic)]
+pub fn rational_function<T, const N: usize, const D: usize>(
+    x: T,
+    numerator: [T; N],
+    denominator: [T; D],
+) -> T
+where
+    T: num_traits::Float,
+{
+    numerator
+        .into_iter()
+        .rev()
+        .fold(T::zero(), |acc, n| acc * x + n)
+        / denominator
+            .into_iter()
+            .rev()
+            .fold(T::zero(), |acc, d| acc * x + d)
 }
